apps/ai_worker/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.1.4
cffi==2.0.0
dnspython==2.8.0
minio==7.2.20
pika==1.3.2
pycparser==3.0
pycryptodome==3.23.0
pymongo==4.16.0
python-dotenv==1.2.1
typing_extensions==4.15.0
urllib3==2.6.3
